<feat> started to separate widgets
