+++ frontend/donutlista/lib/settingspage.dart
@@ -1,0 +1,5 @@
+//TODO: Remove unused imports
+// ignore: unused_import
+import 'package:flutter/material.dart';
+// ignore: unused_import
+import 'package:hexcolor/hexcolor.dart';
