+++ frontend/donutlista/lib/userpage.dart
@@ -1,0 +1,5 @@
+//TODO: Remove unused imports
+// ignore_for_file: unused_import
+
+import 'package:flutter/material.dart';
+import 'package:hexcolor/hexcolor.dart';
